Add for a "goto slide" when clicking the current slide number in the presenter window
Major changes were necessary: there are now much less variables in
pympress.Presenter, which needs access to pympress.Document (esp. to its
"goto()" method).

It is disturbing that the Document, Presenter and Content classes are so tightly
related to each other (GUI management code in Document ?!?). Maybe it's time for
a complete redesign of pympress...