Cache for rendering pages
This stores pixbufs of already rendered pages in memory. This way it is not
necessary to re-render a page to a Cairo surface (which is slow) when a page is
displayed a second time: it is just a matter of copying a bitmap, which
is *much* faster, at the cost of a higher memory consumption.