Modified the following sources to display a dual-screen PDF file that contain talking notes for presenter. To maximize the influence on the existing codes, all modifications are made independently as much as possible.
The Beamer LaTeX package provides the functionality to support dual-screen PDF
file containing notes for presenter, but current the pympress cannot
appropriately display this kind of PDF slides. The main idea to do so is, the
PDF pages should be automatically split to two parts, the contents (left) and
the notes (right), then display the contents in the "Content Window" and the
notes in the "Current Slide" window. Let the "Next Slide" window display the
next content slide as well, or it could be hidden.

The clue of modifying the sources is to add a FLAG parameter to identify the
display mode for the aforementioned kind of PDF file. Details can be confirmed
in the following source files.

	modified:   pympress/document.py
	modified:   pympress/pixbufcache.py
	modified:   pympress/ui.py