Pympress can now open and close documents (fixes #8)
This is done with a fake Document class (and a fake Page class) to be
used as placeholders when we have no valid document open.

A can_render() function has been added to Page to avoid calling an empty
render, hence no drawing on current slide instead of an uglier black rectangle.