Replace setup with __init__ in UI's helper classes
Requires to instantiate helper classes after builder.load_ui() is
called, thus corresponding attributes are initialized to None.

We can do this because we reversed the introspection-loading in builder,
instead of listing attr with None value and getting the corresponding
objects, we now list the built objects and look for corresponding
None-valued attributes.