Refactor zoom into its own standalone class
Slightly clearer than entangled with scribble. Only problem, there is a
circular dependency to setup callbacks, between highlight and zoom. Zoom
requires a callback to toggle the sensitivity of a button in the
scribble pane. Solved by delaying the callback setup. But wouldn't it be
better with some notion of the "unzoom" action at UI level, and
buttons/menu items' sensitivities reflecting the action's availability ?