Fix #100: improve error message for missing pygobject
Should guide users to have a successfull install. pygobject is kept as a
runtime dependency rather than install-time dependency to incite user to
get it through their package manager rather than compiling it with pip.

The error message tries to be complete yet brief, and includes a link to
the dependencies section of the README.

Also remove last remaining gi import in util.