Video support: only show gstreamer output after player is initialized
When having one (or more) videos on a slide there is the following
effect:

* First time opening this slide: 1+ video frames are black; then the
  video starts.
* Second time opening this slide: Video seems to start on second 2 or 3
  and then suddenly jumps backwards.

It seems like this is caused by showing the video output before the
player is fully initialized. This commit only shows the output after
the video is initialized and avoid both effects.

Tested on Ubuntu 20.04 with LaTeX Beamer slides.