Only save window size and position if it is not maximised
Same with fullscreened. Will hopefully decrease issues with windows
being not-maximised but still the size of the monitor and thus
impossible to grab (or even see for the content window).

Note that on wayland we have no control over position: get_position()
returns (0, 0) and move() called from parse_geometry() is nop.

Make sure we queue a draw on the content area when it is resized too.
This might not always be done if there are no slides.